+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Tweak wording in CLI help.
 - Remove extra blank lines between `find_package()` calls in generated CMake files.
+- MSBuild imports related to Qt/MsBuild and Conan no longer generate warnings.
 
 ### Fixed